CRUD funcional, incluyendo aplicar y desaplicar

+++ src/main/java/com/pruebaSpring/tienda/services/PrendaService.java
@@ -5,24 +5,41 @@ import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
-import com.pruebaSpring.tienda.constants.Constants;
+import com.pruebaSpring.tienda.constants.StringConstants;
 import com.pruebaSpring.tienda.models.PrendaModel;
 import com.pruebaSpring.tienda.repositories.PrendaRepository;
 
 @Service
+@Transactional(rollbackFor = Exception.class)
 public class PrendaService {
 
     @Autowired
     PrendaRepository prendaRepository;
 
-    public ArrayList<PrendaModel> obtenerPrendas() {
+    // GET ALL
+    @Transactional(readOnly = true)
+    public ArrayList<PrendaModel> getAllPrenda() {
         return (ArrayList<PrendaModel>) prendaRepository.findAll();
     }
 
-    public PrendaModel guardarPrenda(PrendaModel prenda) {
+    // GET BY ID
+    @Transactional(readOnly = true)
+    public Optional<PrendaModel> getByIdPrenda(String id) {
+        return prendaRepository.findById(id);
+    }
+
+    // POST / PUT (NO COMPROBATIONS - USED WHEN APPLYING DISCOUNT)
+    // AVOID SET PRICE OF NORMAL POST / PUT
+    public PrendaModel ppPrecioPrenda(PrendaModel prenda) {
+        return prendaRepository.save(prenda);
+    }
+
+    // POST / PUT (WITH COMPROBATIONS - USED NORMALLY)
+    public PrendaModel ppComprobationsPrenda(PrendaModel prenda) {
         // Si no coincide con el patrón que no se pueda guardar
-        if (!prenda.getReferencia().matches(Constants.regexRef) || prenda.getPrecio() < 0) {
+        if (!prenda.getReferencia().matches(StringConstants.regexRef) || prenda.getPrecio() < 0) {
             // No se realiza el POST/PUT
             return null;
         }
@@ -32,12 +49,18 @@ public class PrendaService {
         return prendaRepository.save(prenda);
     }
 
-    public Optional<PrendaModel> obtenerPorId(String id) {
-        return prendaRepository.findById(id);
-    }
-
-    public boolean eliminarPrenda(String id) {
+    // DELETE
+    public boolean deletePrenda(String id) {
+        // Intentamos obtener la prenda pedida
+        Optional<PrendaModel> prenda = prendaRepository.findById(id);
+        // Si no se encuentra, no seguimos
+        if (!prenda.isPresent()) {
+            return false;
+        }
         try {
+            // La prenda existe, por lo que borramos sus promociones (tabla intermedia) ->
+            prenda.get().removePromociones();
+            // -> y posteriormente borramos a la misma prenda
             prendaRepository.deleteById(id);
             return true;
         } catch (Exception err) {
